Update TravisCI, UI tests, and marathon data.json, LICENSE year
