+++ .travis.yml
@@ -1,4 +1,6 @@
 language: ruby
+sudo: false
+cache: bundler
 
 branches:
   except:
@@ -6,14 +8,18 @@ branches:
 
 rvm:
   - 2.0.0
+  - 2.4.0
+  - ruby-head
+
+bundler_args: --without production
 
 before_install:
-  - pwd
+  - rm ${BUNDLE_GEMFILE}.lock
 
   - sudo apt-get update
   - sudo apt-get install chromium-browser p7zip-full
 
-  - wget -N http://chromedriver.storage.googleapis.com/2.9/chromedriver_linux64.zip -P ~/Downloads
+  - wget -N https://chromedriver.storage.googleapis.com/2.27/chromedriver_linux64.zip -P ~/Downloads
   - 7z e ~/Downloads/chromedriver_linux64.zip -y
   - ls
   - chmod +x chromedriver
@@ -22,20 +28,29 @@ before_install:
   - sudo ln -s /usr/local/share/chromedriver /usr/bin/chromedriver
   - which chromedriver
 
+  - wget -N https://github.com/mozilla/geckodriver/releases/download/v0.11.1/geckodriver-v0.11.1-linux64.tar.gz -P ~/Downloads
+  - 7z e ~/Downloads/geckodriver-v0.11.1-linux64.tar.gz -y
+  - ls
+  - chmod +x geckodriver
+  - sudo mv -f geckodriver /usr/local/share/geckodriver
+  - sudo ln -s /usr/local/share/geckodriver /usr/local/bin/geckodriver
+  - sudo ln -s /usr/local/share/geckodriver /usr/bin/geckodriver
+  - which chromedriver
+
   - firefox --version
   - phantomjs --version
-
-  - gem update --system
-  - gem --version
-  - gem update
 
 before_script:
   - "export DISPLAY=:99.0"
   - "sh -e /etc/init.d/xvfb start"
   - sleep 3 # give xvfb some time to start
 
-  - rake travis
+  - bundle exec rake travis
 
 script:
-  - rake test:posts
-  - rake test:ui
+  #- bundle exec rake test:posts
+  - cd _tests
+  - cucumber
+
+notifications:
+  email: false
