--- _rake/features/support/pages/list_view.rb
@@ -1,9 +1,0 @@
-require './_rake/features/support/pages/base_page.rb'
-
-module Pages
-	class ListView < BasePage
-		def posts
-			return body_content.find_elements(:class => 'posts')
-		end
-	end
-end
