+++ _tests/features/blog_body.feature
@@ -5,7 +5,7 @@ Feature: Test blog page body
 
 	Scenario Outline: Body's title header should be displayed
 		When I set window width to <width>px
-		Then I should see title header "Top Posts"
+		Then I should see title header "All Posts"
 
 		Examples:
 			| width |
@@ -17,7 +17,7 @@ Feature: Test blog page body
 
 	Scenario Outline: Body should list posts
 		When I set window width to <width>px
-		Then I should have posts grouped by "2013, 2014, 2015"
+		Then I should have posts grouped by "2013, 2014, 2015, 2016"
 		And I should have clickable post links
 
 		Examples:
