+++ _tests/features/support/pages/list_view.rb
@@ -1,0 +1,9 @@
+require_relative './base_page.rb'
+
+module Pages
+  class ListView < BasePage
+    def posts
+      body_content.find_elements(class: 'posts')
+    end
+  end
+end
